[AI-Generated] Create Chat Session Management Service
Task ID: task-014

Auto-generated by AI Agent workflow

+++ backend/src/models/chat.py
@@ -150,7 +150,7 @@ class ChatMessage(Base):
 
     # Primary key
     id: Mapped[int] = mapped_column(
-        BigInteger,
+        Integer,
         primary_key=True,
         autoincrement=True,
         comment="Unique message identifier"
